Initial commit of IMPLEMENTATION.md
To accelerate this project, we'll begin with a set of tutorials which
introduce the concepts with real-world use. Just this will take several
hours for someone who hasn't done this before. Then the project will be
to bring it all together: the implementation will do this, hopefully
also including CI/CD.