Update README.md to remove firecracker section  (#83)
Update README.md

I'm removing this section for now - while this is doable if you have access to a linux box it turns out that running firecracker on apple silicon (which I believe is what students are using) is not currently feasible because none of the virtualisation tools seem to support nested virtualisation on apple silicon yet. 
Can be done on an EC2 metal instance but that's really expensive and probably not worth it.