Add flattened version of the workbook
For getting an overview of what the course covered, this is easier to
scan at a glance than having to click through each sprint on the
existing page.

I can't work out how to make this automatic for each entry in the
taxonomy, but this is useful enough for my purposes for now.